Removed the 64 int array and replaced with a PieceSet object
Now a bunch of stuffs are broken, I need to debug
Also many functions can be updated to use the piece masks now

+++ src/chess/model/board_geometry.rs
@@ -14,7 +14,7 @@ use std::cmp::max;
 /// The number of moves the king needs to do to reach the destination. With valid
 /// square indexes, it will be in [0..7]
 ///
-pub fn get_king_distance(king_position: usize, destination: usize) -> usize {
+pub fn get_king_distance(king_position: u8, destination: u8) -> u8 {
   let (sf, sr) = Board::index_to_fr(king_position);
   let (df, dr) = Board::index_to_fr(destination);
